address AI PR comment

+++ src/serializers/decoders/prev/v18/decode_udf.c
@@ -24,7 +24,7 @@ void AUXLoadUDF_v18
 	// read the library count
 	uint64_t n = RedisModule_LoadUnsigned (io) ;
 
-	for (unsigned int i = 0; i < n; i++) {
+	for (uint64_t i = 0; i < n; i++) {
 		size_t lib_len ;
 		size_t script_len ;
 		const char *lib    = RedisModule_LoadStringBuffer (io, &lib_len) ;
